requirements.txt
faiss_cpu==1.7.3
Flask==2.2.2
numpy==1.24.1
Pillow==9.4.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
Werkzeug==2.2.2